requirements.txt
fastapi
uvicorn
requests
python-dotenv
spotipy
PyQt5
websockets
websockets-client
asyncio
pyler